pyproject.toml
[project]
name = "mrboost"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Chunxu Guo", email = "chunxu@wustl.edu" }
]
dependencies = [
    "scipy>=1.13.0",
    "numpy>=1.26.4",
    "einops>=0.8.0",
    "tqdm>=4.66.2",
    "h5py>=3.11.0",
    "nibabel>=5.2.1",
    "pydicom>=2.4.4",
    "zarr>=2.17.2",
    "torchkbnufft>=1.4.0",
    "xarray>=2024.3.0",
    "jax[cpu]>=0.4.26",
    "fire>=0.6.0",
    "torch>=2.3.0",
    "pytorch-finufft @ git+https://github.com/flatironinstitute/pytorch-finufft.git",
    "matplotlib>=3.9.0",
    "plum-dispatch>=2.3.6",
    "cufinufft>=2.2.0",
    "einx>=0.3.0",
    "jaxtyping>=0.2.33",
    "twixtools @ git+https://github.com/pehses/twixtools.git",
    "setuptools>=74.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.5",
    "icecream>=2.1.3",
    "monai>=1.3.0",
    "lightning>=2.2.3",
    "torchopt>=0.7.3",
    "wandb>=0.16.6",
    "beartype>=0.18.5",
    "scikit-learn>=1.4.2",
    "rich>=13.7.1",
    "dask[distributed]>=2024.4.2",
    "torch-dct>=0.1.6",
    "imageio>=2.34.1",
    "jsonargparse[signatures]>=4.32.1",
    "omegaconf>=2.3.0",
    "optree>=0.11.0",
    "mri-nufft>=0.9.1",
    "pytest>=8.2.2",
    "pytorch-minimize>=0.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mrboost"]
